pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-manual-rag-assistant"
version = "0.1.0"
description = "政企运维知识库问答助手"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "python-docx>=1.1.2,<2.0.0",
  "openpyxl>=3.1.5,<4.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
